Make it a view.
